+++ src/main/java/com/google/cloud/tools/gradle/appengine/standard/AppEngineStandardPlugin.java
@@ -97,7 +97,8 @@ public class AppEngineStandardPlugin implements Plugin<Project> {
           // tools extension required to initialize cloudSdkOperations
           ToolsExtension tools = appengineExtension.getTools();
           try {
-            cloudSdkOperations = new CloudSdkOperations(tools.getCloudSdkHome(), null);
+            cloudSdkOperations =
+                new CloudSdkOperations(tools.getCloudSdkHome(), null, tools.getVerbosity());
           } catch (CloudSdkNotFoundException ex) {
             // this should be caught in AppEngineCorePluginConfig before it can ever reach here.
             throw new GradleException("Could not find CloudSDK: ", ex);